+++ src/components/commandHub/StarButton.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useCallback } from 'react'
-import { useSection, CACHE_SECTIONS } from '@/hooks/useEnhancedSetting'
+import { useSection } from '@/hooks/useSetting'
+import { CACHE_SECTIONS } from '@/services/settingsCache'
 import { sendEvent, ANALYTICS_EVENTS } from '@/services/analytics'
 import { SCREEN } from '@/const'
 import { useDetectUrlChanged } from '@/hooks/useDetectUrlChanged'
